Implemented friend function
When you select a username, it bolds their messages if you're adding them as a friend.
If they are already a friend, you unfriend them and it unfriends their messages.
Some bugs when usernames start with a number or contain strange characters because of how
we inserted the username as the class of every message text span element.